refactor: move movie package
