+++ src/test/java/com/example/cleancoder/functionstructure/movie/VideoStoreTest.java
@@ -1,4 +1,4 @@
-package com.example.cleancoder.functionstructure;
+package com.example.cleancoder.functionstructure.movie;
 
 import org.assertj.core.data.Offset;
 import org.junit.Before;
